MedHap
MedHap is a medical application that models the patient's skin textural
abnormalities on Doctor's digital devices. Developed at Cal Hacks 3.0 on
November 11th-13th leading a team of 5.
Brief Explanation: In this the patient register's on the application as
a patient. After this the patient takes a picture of the skin allergies
such as moles, acne, rash etc.. This image is then uploaded to the
server.. The doctor then reviews these images using the powerful Tanvas
SDK integrated.. Using the haptic feedback the doctor can feel the rash
or mole and let the patient know of his condition..